more coverage (#13)

+++ scalars.go
@@ -23,6 +23,7 @@ func parseDateTime(v interface{}) interface{} {
 	return nil
 }
 
+// DateTimeType provides a DateTime implementation that serializing to and from RFC-3339 datetimes.
 var DateTimeType = &graphql.ScalarType{
 	Name:        "DateTime",
 	Description: "DateTime represents an RFC-3339 datetime.",
@@ -45,6 +46,8 @@ var DateTimeType = &graphql.ScalarType{
 	},
 }
 
+// NonZeroDateTime returns a field definition that resolves to the value of the field with the given
+// name. If the field's value is the zero time, the field resolves to nil instead.
 func NonZeroDateTime(fieldName string) *graphql.FieldDefinition {
 	return &graphql.FieldDefinition{
 		Type: DateTimeType,
@@ -83,30 +86,32 @@ func coerceLongInt(v interface{}) interface{} {
 		return int64(v)
 	case int64:
 		if v >= minSafeInteger && v <= maxSafeInteger {
-			return v
+			return int64(v)
 		}
 	case uint64:
 		if v <= maxSafeInteger {
-			return v
+			return int64(v)
 		}
 	case int:
 		if v >= minSafeInteger && v <= maxSafeInteger {
-			return v
+			return int64(v)
 		}
 	case uint:
 		if v <= maxSafeInteger {
-			return v
+			return int64(v)
 		}
 	case float32:
 		return coerceLongInt(float64(v))
 	case float64:
 		if n := math.Trunc(v); n == v && n >= minSafeInteger && n <= maxSafeInteger {
-			return int(n)
+			return int64(n)
 		}
 	}
 	return nil
 }
 
+// LongIntType provides a scalar implementation for integers that may be larger than 32 bits, but
+// can still be represented by JavaScript numbers.
 var LongIntType = &graphql.ScalarType{
 	Name:        "LongInt",
 	Description: "LongInt represents a signed integer that may be longer than 32 bits, but still within JavaScript / IEEE-654's \"safe\" range.",
@@ -114,7 +119,7 @@ var LongIntType = &graphql.ScalarType{
 		switch v := v.(type) {
 		case *ast.IntValue:
 			if n, err := strconv.ParseInt(v.Value, 10, 64); err == nil && n >= minSafeInteger && n <= maxSafeInteger {
-				return int(n)
+				return n
 			}
 		}
 		return nil
